lib/netext/grpcext: Split the gRPC module
It splits the gRPC business logic from the js/k6/grpc module to a netext
package.
It will be easier to extend the networking+metric features for gRPC in
isolated environment and/or create extension not strictly depending on
the js feature.
