+++ js/modules/k6/grpc/client.go
@@ -1,94 +1,37 @@
-/*
- *
- * k6 - a next-generation load testing tool
- * Copyright (C) 2020 Load Impact
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as
- * published by the Free Software Foundation, either version 3 of the
- * License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- *
- */
-
 package grpc
 
 import (
 	"context"
-	"encoding/json"
 	"errors"
 	"fmt"
 	"io"
-	"net"
-	"strconv"
 	"strings"
 	"time"
+
+	"go.k6.io/k6/js/common"
+	"go.k6.io/k6/js/modules"
+	"go.k6.io/k6/lib/netext/grpcext"
+	"go.k6.io/k6/lib/types"
+	"go.k6.io/k6/metrics"
 
 	"github.com/dop251/goja"
 	"github.com/jhump/protoreflect/desc"
 	"github.com/jhump/protoreflect/desc/protoparse"
-	"github.com/sirupsen/logrus"
 	"google.golang.org/grpc"
-	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/credentials"
 	"google.golang.org/grpc/credentials/insecure"
 	"google.golang.org/grpc/metadata"
-	grpcstats "google.golang.org/grpc/stats"
-	"google.golang.org/grpc/status"
-	"google.golang.org/protobuf/encoding/protojson"
-	"google.golang.org/protobuf/encoding/prototext"
-	"google.golang.org/protobuf/proto"
 	"google.golang.org/protobuf/reflect/protodesc"
 	"google.golang.org/protobuf/reflect/protoreflect"
 	"google.golang.org/protobuf/types/descriptorpb"
-	"google.golang.org/protobuf/types/dynamicpb"
-
-	//nolint: staticcheck
-	protoV1 "github.com/golang/protobuf/proto"
-
-	"go.k6.io/k6/js/common"
-	"go.k6.io/k6/js/modules"
-	"go.k6.io/k6/lib/types"
-	"go.k6.io/k6/metrics"
-	reflectpb "google.golang.org/grpc/reflection/grpc_reflection_v1alpha"
-)
-
-//nolint: lll
-var (
-	errInvokeRPCInInitContext = common.NewInitContextError("invoking RPC methods in the init context is not supported")
-	errConnectInInitContext   = common.NewInitContextError("connecting to a gRPC server in the init context is not supported")
 )
 
 // Client represents a gRPC client that can be used to make RPC requests
 type Client struct {
 	mds  map[string]protoreflect.MethodDescriptor
-	conn *grpc.ClientConn
-
-	vu modules.VU
-}
-
-// MethodInfo holds information on any parsed method descriptors that can be used by the goja VM
-type MethodInfo struct {
-	grpc.MethodInfo `json:"-" js:"-"`
-	Package         string
-	Service         string
-	FullMethod      string
-}
-
-// Response is a gRPC response that can be used by the goja VM
-type Response struct {
-	Status   codes.Code
-	Message  interface{}
-	Headers  map[string][]string
-	Trailers map[string][]string
-	Error    interface{}
+	conn *grpcext.Conn
+	vu   modules.VU
+	addr string
 }
 
 // Load will parse the given proto files and make the file descriptors available to request.
@@ -134,7 +77,7 @@ func (c *Client) Load(importPaths []string, filenames ...string) ([]MethodInfo, 
 func (c *Client) Connect(addr string, params map[string]interface{}) (bool, error) {
 	state := c.vu.State()
 	if state == nil {
-		return false, errConnectInInitContext
+		return false, common.NewInitContextError("connecting to a gRPC server in the init context is not supported")
 	}
 
 	p, err := c.parseConnectParams(params)
@@ -142,30 +85,28 @@ func (c *Client) Connect(addr string, params map[string]interface{}) (bool, erro
 		return false, err
 	}
 
-	opts := make([]grpc.DialOption, 0, 2)
+	opts := grpcext.DefaultOptions(c.vu)
 
+	var tcred credentials.TransportCredentials
 	if !p.IsPlaintext {
 		tlsCfg := state.TLSConfig.Clone()
 		tlsCfg.NextProtos = []string{"h2"}
 
 		// TODO(rogchap): Would be good to add support for custom RootCAs (self signed)
-
-		opts = append(opts, grpc.WithTransportCredentials(credentials.NewTLS(tlsCfg)))
+		tcred = credentials.NewTLS(tlsCfg)
 	} else {
-		opts = append(opts, grpc.WithTransportCredentials(insecure.NewCredentials()))
+		tcred = insecure.NewCredentials()
 	}
+	opts = append(opts, grpc.WithTransportCredentials(tcred))
 
 	if ua := state.Options.UserAgent; ua.Valid {
 		opts = append(opts, grpc.WithUserAgent(ua.ValueOrZero()))
 	}
 
-	dialer := func(ctx context.Context, addr string) (net.Conn, error) {
-		return state.Dialer.DialContext(ctx, "tcp", addr)
-	}
-	opts = append(opts, grpc.WithContextDialer(dialer))
-
 	ctx, cancel := context.WithTimeout(c.vu.Context(), p.Timeout)
 	defer cancel()
+
+	c.addr = addr
 
 	err = c.dial(ctx, addr, p.UseReflectionProtocol, opts...)
 	return err != nil, err
@@ -176,11 +117,10 @@ func (c *Client) Invoke(
 	method string,
 	req goja.Value,
 	params map[string]interface{},
-) (*Response, error) {
-	rt := c.vu.Runtime()
+) (*grpcext.Response, error) {
 	state := c.vu.State()
 	if state == nil {
-		return nil, errInvokeRPCInInitContext
+		return nil, common.NewInitContextError("invoking RPC methods in the init context is not supported")
 	}
 	if c.conn == nil {
 		return nil, errors.New("no gRPC connection, you must call connect first")
@@ -191,8 +131,8 @@ func (c *Client) Invoke(
 	if method[0] != '/' {
 		method = "/" + method
 	}
-	md := c.mds[method]
-	if md == nil {
+	methodDesc := c.mds[method]
+	if methodDesc == nil {
 		return nil, fmt.Errorf("method %q not found in file descriptors", method)
 	}
 
@@ -201,10 +141,18 @@ func (c *Client) Invoke(
 		return nil, err
 	}
 
-	ctx := metadata.NewOutgoingContext(c.vu.Context(), metadata.New(nil))
-	for param, strval := range p.Metadata {
-		ctx = metadata.AppendToOutgoingContext(ctx, param, strval)
+	b, err := req.ToObject(c.vu.Runtime()).MarshalJSON()
+	if err != nil {
+		return nil, fmt.Errorf("unable to serialise request object: %w", err)
 	}
+
+	md := metadata.New(nil)
+	for param, strval := range p.Metadata {
+		md.Append(param, strval)
+	}
+
+	ctx, cancel := context.WithTimeout(c.vu.Context(), p.Timeout)
+	defer cancel()
 
 	tags := state.CloneTags()
 	for k, v := range p.Tags {
@@ -212,7 +160,7 @@ func (c *Client) Invoke(
 	}
 
 	if state.Options.SystemTags.Has(metrics.TagURL) {
-		tags["url"] = fmt.Sprintf("%s%s", c.conn.Target(), method)
+		tags["url"] = fmt.Sprintf("%s%s", c.addr, method)
 	}
 	parts := strings.Split(method[1:], "/")
 	if state.Options.SystemTags.Has(metrics.TagService) {
@@ -227,80 +175,32 @@ func (c *Client) Invoke(
 		tags["name"] = method
 	}
 
-	ctx = withTags(ctx, tags)
-
-	reqdm := dynamicpb.NewMessage(md.Input())
-	{
-		b, err := req.ToObject(rt).MarshalJSON()
-		if err != nil {
-			return nil, fmt.Errorf("unable to serialise request object: %w", err)
-		}
-		if err := protojson.Unmarshal(b, reqdm); err != nil {
-			return nil, fmt.Errorf("unable to serialise request object to protocol buffer: %w", err)
-		}
+	reqmsg := grpcext.Request{
+		MethodDescriptor: methodDesc,
+		Message:          b,
+		Tags:             tags,
 	}
 
-	reqCtx, cancel := context.WithTimeout(ctx, p.Timeout)
-	defer cancel()
-
-	resp := dynamicpb.NewMessage(md.Output())
-	header, trailer := metadata.New(nil), metadata.New(nil)
-	err = c.conn.Invoke(reqCtx, method, reqdm, resp, grpc.Header(&header), grpc.Trailer(&trailer))
-
-	var response Response
-	response.Headers = header
-	response.Trailers = trailer
-
-	marshaler := protojson.MarshalOptions{EmitUnpopulated: true}
-
-	if err != nil {
-		sterr := status.Convert(err)
-		response.Status = sterr.Code()
-
-		// (rogchap) when you access a JSON property in goja, you are actually accessing the underling
-		// Go type (struct, map, slice etc); because these are dynamic messages the Unmarshaled JSON does
-		// not map back to a "real" field or value (as a normal Go type would). If we don't marshal and then
-		// unmarshal back to a map, you will get "undefined" when accessing JSON properties, even when
-		// JSON.Stringify() shows the object to be correctly present.
-
-		raw, _ := marshaler.Marshal(sterr.Proto())
-		errMsg := make(map[string]interface{})
-		_ = json.Unmarshal(raw, &errMsg)
-		response.Error = errMsg
-	}
-
-	if resp != nil {
-		// (rogchap) there is a lot of marshaling/unmarshaling here, but if we just pass the dynamic message
-		// the default Marshaller would be used, which would strip any zero/default values from the JSON.
-		// eg. given this message:
-		// message Point {
-		//    double x = 1;
-		// 	  double y = 2;
-		// 	  double z = 3;
-		// }
-		// and a value like this:
-		// msg := Point{X: 6, Y: 4, Z: 0}
-		// would result in JSON output:
-		// {"x":6,"y":4}
-		// rather than the desired:
-		// {"x":6,"y":4,"z":0}
-		raw, _ := marshaler.Marshal(resp)
-		msg := make(map[string]interface{})
-		_ = json.Unmarshal(raw, &msg)
-		response.Message = msg
-	}
-	return &response, nil
+	return c.conn.Invoke(ctx, method, md, reqmsg)
 }
 
 // Close will close the client gRPC connection
 func (c *Client) Close() error {
-	if c == nil || c.conn == nil {
+	if c.conn == nil {
 		return nil
 	}
 	err := c.conn.Close()
 	c.conn = nil
 
 	return err
+}
+
+// MethodInfo holds information on any parsed method descriptors that can be used by the goja VM
+type MethodInfo struct {
+	Package         string
+	Service         string
+	FullMethod      string
+	grpc.MethodInfo `json:"-" js:"-"`
 }
 
 func (c *Client) convertToMethodInfo(fdset *descriptorpb.FileDescriptorSet) ([]MethodInfo, error) {
@@ -353,16 +253,8 @@ func (c *Client) dial(
 	reflect bool,
 	options ...grpc.DialOption,
 ) error {
-	opts := []grpc.DialOption{
-		grpc.WithBlock(),
-		grpc.FailOnNonTempDialError(true),
-		grpc.WithStatsHandler(statsHandler{vu: c.vu}),
-		grpc.WithReturnConnectionError(),
-	}
-	opts = append(opts, options...)
-
 	var err error
-	c.conn, err = grpc.DialContext(ctx, addr, opts...)
+	c.conn, err = grpcext.Dial(ctx, addr, options...)
 	if err != nil {
 		return err
 	}
@@ -370,32 +262,13 @@ func (c *Client) dial(
 	if !reflect {
 		return nil
 	}
-
-	return c.reflect(ctx)
-}
-
-// reflect will use the grpc reflection api to make the file descriptors available to request.
-// It is called in the connect function the first time the Client.Connect function is called.
-func (c *Client) reflect(ctx context.Context) error {
-	client := reflectpb.NewServerReflectionClient(c.conn)
-	methodClient, err := client.ServerReflectionInfo(ctx)
+	rc, err := c.conn.ReflectionClient()
 	if err != nil {
-		return fmt.Errorf("can't get server info: %w", err)
+		return err
 	}
-	req := &reflectpb.ServerReflectionRequest{
-		MessageRequest: &reflectpb.ServerReflectionRequest_ListServices{},
-	}
-	resp, err := sendReceive(methodClient, req)
+	fdset, err := rc.Reflect(ctx)
 	if err != nil {
-		return fmt.Errorf("can't list services: %w", err)
-	}
-	listResp := resp.GetListServicesResponse()
-	if listResp == nil {
-		return fmt.Errorf("can't list services, nil response")
-	}
-	fdset, err := resolveServiceFileDescriptors(methodClient, listResp)
-	if err != nil {
-		return fmt.Errorf("can't resolve services' file descriptors: %w", err)
+		return err
 	}
 	_, err = c.convertToMethodInfo(fdset)
 	if err != nil {
@@ -501,140 +374,6 @@ func (c *Client) parseConnectParams(raw map[string]interface{}) (connectParams, 
 	return params, nil
 }
 
-type statsHandler struct {
-	vu modules.VU
-}
-
-// TagConn implements the grpcstats.Handler interface
-func (statsHandler) TagConn(ctx context.Context, _ *grpcstats.ConnTagInfo) context.Context {
-	// noop
-	return ctx
-}
-
-// HandleConn implements the grpcstats.Handler interface
-func (statsHandler) HandleConn(context.Context, grpcstats.ConnStats) {
-	// noop
-}
-
-// TagRPC implements the grpcstats.Handler interface
-func (statsHandler) TagRPC(ctx context.Context, _ *grpcstats.RPCTagInfo) context.Context {
-	// noop
-	return ctx
-}
-
-// HandleRPC implements the grpcstats.Handler interface
-func (h statsHandler) HandleRPC(ctx context.Context, stat grpcstats.RPCStats) {
-	state := h.vu.State()
-	tags := getTags(ctx)
-	switch s := stat.(type) {
-	case *grpcstats.OutHeader:
-		if state.Options.SystemTags.Has(metrics.TagIP) && s.RemoteAddr != nil {
-			if ip, _, err := net.SplitHostPort(s.RemoteAddr.String()); err == nil {
-				tags["ip"] = ip
-			}
-		}
-	case *grpcstats.End:
-		if state.Options.SystemTags.Has(metrics.TagStatus) {
-			tags["status"] = strconv.Itoa(int(status.Code(s.Error)))
-		}
-
-		mTags := map[string]string(tags)
-		sampleTags := metrics.IntoSampleTags(&mTags)
-		metrics.PushIfNotDone(ctx, state.Samples, metrics.ConnectedSamples{
-			Samples: []metrics.Sample{
-				{
-					Metric: state.BuiltinMetrics.GRPCReqDuration,
-					Tags:   sampleTags,
-					Value:  metrics.D(s.EndTime.Sub(s.BeginTime)),
-					Time:   s.EndTime,
-				},
-			},
-		})
-	}
-
-	// (rogchap) Re-using --http-debug flag as gRPC is technically still HTTP
-	if state.Options.HTTPDebug.String != "" {
-		logger := state.Logger.WithField("source", "http-debug")
-		httpDebugOption := state.Options.HTTPDebug.String
-		debugStat(stat, logger, httpDebugOption)
-	}
-}
-
-// sendReceiver is a smaller interface for decoupling
-// from `reflectpb.ServerReflection_ServerReflectionInfoClient`,
-// that has the dependency from `grpc.ClientStream`,
-// which is too much in the case the requirement is to just make a reflection's request.
-// It makes the API more restricted and with a controlled surface,
-// in this way the testing should be easier also.
-type sendReceiver interface {
-	Send(*reflectpb.ServerReflectionRequest) error
-	Recv() (*reflectpb.ServerReflectionResponse, error)
-}
-
-// sendReceive sends a request to a reflection client and,
-// receives a response.
-func sendReceive(
-	client sendReceiver,
-	req *reflectpb.ServerReflectionRequest,
-) (*reflectpb.ServerReflectionResponse, error) {
-	if err := client.Send(req); err != nil {
-		return nil, fmt.Errorf("can't send request: %w", err)
-	}
-	resp, err := client.Recv()
-	if err != nil {
-		return nil, fmt.Errorf("can't receive response: %w", err)
-	}
-	return resp, nil
-}
-
-type fileDescriptorLookupKey struct {
-	Package string
-	Name    string
-}
-
-func resolveServiceFileDescriptors(
-	client sendReceiver,
-	res *reflectpb.ListServiceResponse,
-) (*descriptorpb.FileDescriptorSet, error) {
-	services := res.GetService()
-	seen := make(map[fileDescriptorLookupKey]bool, len(services))
-	fdset := &descriptorpb.FileDescriptorSet{
-		File: make([]*descriptorpb.FileDescriptorProto, 0, len(services)),
-	}
-
-	for _, service := range services {
-		req := &reflectpb.ServerReflectionRequest{
-			MessageRequest: &reflectpb.ServerReflectionRequest_FileContainingSymbol{
-				FileContainingSymbol: service.GetName(),
-			},
-		}
-		resp, err := sendReceive(client, req)
-		if err != nil {
-			return nil, fmt.Errorf("can't get method on service %q: %w", service, err)
-		}
-		fdResp := resp.GetFileDescriptorResponse()
-		for _, raw := range fdResp.GetFileDescriptorProto() {
-			var fdp descriptorpb.FileDescriptorProto
-			if err = proto.Unmarshal(raw, &fdp); err != nil {
-				return nil, fmt.Errorf("can't unmarshal proto on service %q: %w", service, err)
-			}
-			fdkey := fileDescriptorLookupKey{
-				Package: *fdp.Package,
-				Name:    *fdp.Name,
-			}
-			if seen[fdkey] {
-				// When a proto file contains declarations for multiple services
-				// then the same proto file is returned multiple times,
-				// this prevents adding the returned proto file as a duplicate.
-				continue
-			}
-			seen[fdkey] = true
-			fdset.File = append(fdset.File, &fdp)
-		}
-	}
-	return fdset, nil
-}
-
 func walkFileDescriptors(seen map[string]struct{}, fd *desc.FileDescriptor) []*descriptorpb.FileDescriptorProto {
 	fds := []*descriptorpb.FileDescriptorProto{}
 
@@ -650,68 +389,4 @@ func walkFileDescriptors(seen map[string]struct{}, fd *desc.FileDescriptor) []*d
 	}
 
 	return fds
-}
-
-func debugStat(stat grpcstats.RPCStats, logger logrus.FieldLogger, httpDebugOption string) {
-	switch s := stat.(type) {
-	case *grpcstats.OutHeader:
-		logger.Infof("Out Header:\nFull Method: %s\nRemote Address: %s\n%s\n",
-			s.FullMethod, s.RemoteAddr, formatMetadata(s.Header))
-	case *grpcstats.OutTrailer:
-		if len(s.Trailer) > 0 {
-			logger.Infof("Out Trailer:\n%s\n", formatMetadata(s.Trailer))
-		}
-	case *grpcstats.OutPayload:
-		if httpDebugOption == "full" {
-			logger.Infof("Out Payload:\nWire Length: %d\nSent Time: %s\n%s\n\n",
-				s.WireLength, s.SentTime, formatPayload(s.Payload))
-		}
-	case *grpcstats.InHeader:
-		if len(s.Header) > 0 {
-			logger.Infof("In Header:\nWire Length: %d\n%s\n", s.WireLength, formatMetadata(s.Header))
-		}
-	case *grpcstats.InTrailer:
-		if len(s.Trailer) > 0 {
-			logger.Infof("In Trailer:\nWire Length: %d\n%s\n", s.WireLength, formatMetadata(s.Trailer))
-		}
-	case *grpcstats.InPayload:
-		if httpDebugOption == "full" {
-			logger.Infof("In Payload:\nWire Length: %d\nReceived Time: %s\n%s\n\n",
-				s.WireLength, s.RecvTime, formatPayload(s.Payload))
-		}
-	}
-}
-
-func formatMetadata(md metadata.MD) string {
-	var sb strings.Builder
-	for k, v := range md {
-		sb.WriteString(k)
-		sb.WriteString(": ")
-		sb.WriteString(strings.Join(v, ", "))
-		sb.WriteRune('\n')
-	}
-
-	return sb.String()
-}
-
-func formatPayload(payload interface{}) string {
-	msg, ok := payload.(proto.Message)
-	if !ok {
-		// check to see if we are dealing with a APIv1 message
-		msgV1, ok := payload.(protoV1.Message)
-		if !ok {
-			return ""
-		}
-		msg = protoV1.MessageV2(msgV1)
-	}
-
-	marshaler := prototext.MarshalOptions{
-		Multiline: true,
-		Indent:    "  ",
-	}
-	b, err := marshaler.Marshal(msg)
-	if err != nil {
-		return ""
-	}
-	return string(b)
 }
