--- js/modules/k6/grpc/tags.go
@@ -1,39 +1,0 @@
-/*
- *
- * k6 - a next-generation load testing tool
- * Copyright (C) 2020 Load Impact
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as
- * published by the Free Software Foundation, either version 3 of the
- * License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- *
- */
-
-package grpc
-
-import "context"
-
-type ctxKeyTags struct{}
-
-type reqtags map[string]string
-
-func withTags(ctx context.Context, tags reqtags) context.Context {
-	return context.WithValue(ctx, ctxKeyTags{}, tags)
-}
-
-func getTags(ctx context.Context) reqtags {
-	v := ctx.Value(ctxKeyTags{})
-	if v == nil {
-		return make(map[string]string)
-	}
-	return v.(reqtags)
-}
